feat: add Speed, NightVision, AutoArmor modules + GitHub Actions CI

+++ src/main/java/cc/silk/utils/mc/InventoryUtil.java
@@ -10,7 +10,7 @@ import java.util.Objects;
 @UtilityClass
 public final class InventoryUtil implements IMinecraft {
     public static void swapToSlot(Item item) {
-        for (byte i = 0; i < 9; i++) {
+        for (int i = 0; i < 9; i++) {
             assert mc.player != null;
             var stack = mc.player.getInventory().getStack(i);
 
@@ -23,7 +23,7 @@ public final class InventoryUtil implements IMinecraft {
     }
 
     public static boolean hasItem(Item item) {
-        for (byte i = 0; i < Objects.requireNonNull(mc.player).getInventory().size(); i++) {
+        for (int i = 0; i < Objects.requireNonNull(mc.player).getInventory().size(); i++) {
             ItemStack stack = mc.player.getInventory().getStack(i);
             if (stack.getItem() == item) {
                 return true;
@@ -33,7 +33,7 @@ public final class InventoryUtil implements IMinecraft {
     }
 
     public static boolean hasWeapon(Class<? extends Item> weaponClass) {
-        for (byte i = 0; i < Objects.requireNonNull(mc.player).getInventory().size(); i++) {
+        for (int i = 0; i < Objects.requireNonNull(mc.player).getInventory().size(); i++) {
             ItemStack stack = mc.player.getInventory().getStack(i);
             if (weaponClass.isInstance(stack.getItem())) {
                 return true;
@@ -43,7 +43,7 @@ public final class InventoryUtil implements IMinecraft {
     }
 
     public static void swapToWeapon(Class<? extends Item> weaponClass) {
-        for (byte i = 0; i < Objects.requireNonNull(mc.player).getInventory().size(); i++) {
+        for (int i = 0; i < Objects.requireNonNull(mc.player).getInventory().size(); i++) {
             ItemStack stack = mc.player.getInventory().getStack(i);
             if (weaponClass.isInstance(stack.getItem())) {
                 mc.player.getInventory().selectedSlot = i;
